make more analysis submodules and get some tests to work
